Add script that applies chezmoi settings
This script is meant to be used as a `dotfiles.installCommand` in VS Code. We need it, because it's not possible to specify a command with arguments (https://github.com/microsoft/vscode-remote-release/issues/3047).

The chezmoi docs describe how to use it with Dev Containers: https://www.chezmoi.io/user-guide/machines/containers-and-vms/.