Fix warning when chezmoi is not installed
It might be seem strange to have to check for the existence of chezmoi, considering that it is what's used to manage the dotfiles. This could happen if `~/.local/bin` is not in `PATH` (e.g. for Ubuntu login shells as started by WSL, where this location is added to `PATH` after source `~/.bashrc`). It could also happen when using the `--one-shot` option, where all traces of chezmoi are purged after initializing.